Updated structure
Updated the structure by moving the Parser from the Client file to a seperate folder. This will allow both client.py and server.py to parse the commands without replicating the code.

Moreover, added a dummy main file for later use. This will be helpful to navigate between the assignments - instead of memorizing the files requires, it will be possible to run main.py 1, main.py 2 and so on.