Fixed disconnection
The connection with the server is now closed gracefully with the help of socket.shutdown().

Client side is split into two files - client.py and clientOld.py. The former needs to be extended with the code from the latter.

A new file, util.py, was created. It is made for helper functions and such should be extracted from client.py.